Changed the input type to float

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = int(input('Enter your monthly income: '))
-expenses = int(input('Enter your total monthly expenses: '))
+income = float(input('Enter your monthly income: '))
+expenses = float(input('Enter your total monthly expenses: '))
 savings = income - expenses
 i = 0.05
 projected_savings = (savings) * 12 + (savings * 12 * i)
